split ResourceBar and ResourceEditor from ProjectResource
